remove ApiKeys#parseResponse and ApiKeys#parseRequest
