+++ clients/src/test/java/org/apache/kafka/common/requests/ApiVersionsResponseTest.java
@@ -19,10 +19,15 @@ package org.apache.kafka.common.requests;
 
 import org.apache.kafka.common.message.ApiVersionsResponseData.ApiVersionsResponseKey;
 import org.apache.kafka.common.protocol.ApiKeys;
+import org.apache.kafka.common.protocol.CommonFields;
+import org.apache.kafka.common.protocol.types.BoundField;
+import org.apache.kafka.common.protocol.types.Schema;
 import org.junit.Test;
 
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
@@ -67,6 +72,29 @@ public class ApiVersionsResponseTest {
         assertTrue(ApiVersionsResponse.DEFAULT_API_VERSIONS_RESPONSE.data.supportedFeatures().isEmpty());
         assertTrue(ApiVersionsResponse.DEFAULT_API_VERSIONS_RESPONSE.data.finalizedFeatures().isEmpty());
         assertEquals(ApiVersionsResponse.UNKNOWN_FINALIZED_FEATURES_EPOCH, ApiVersionsResponse.DEFAULT_API_VERSIONS_RESPONSE.data.finalizedFeaturesEpoch());
+    }
+
+    /**
+     * All valid client responses which may be throttled should have a field named
+     * 'throttle_time_ms' to return the throttle time to the client. Exclusions are
+     * <ul>
+     *   <li> Cluster actions used only for inter-broker are throttled only if unauthorized
+     *   <li> SASL_HANDSHAKE and SASL_AUTHENTICATE are not throttled when used for authentication
+     *        when a connection is established or for re-authentication thereafter; these requests
+     *        return an error response that may be throttled if they are sent otherwise.
+     * </ul>
+     */
+    @Test
+    public void testResponseThrottleTime() {
+        List<ApiKeys> authenticationKeys = Arrays.asList(ApiKeys.SASL_HANDSHAKE, ApiKeys.SASL_AUTHENTICATE);
+        for (ApiKeys apiKey: ApiKeys.values()) {
+            Schema responseSchema = apiKey.responseSchemas[apiKey.latestVersion()];
+            BoundField throttleTimeField = responseSchema.get(CommonFields.THROTTLE_TIME_MS.name);
+            if (apiKey.clusterAction || authenticationKeys.contains(apiKey))
+                assertNull("Unexpected throttle time field: " + apiKey, throttleTimeField);
+            else
+                assertNotNull("Throttle time field missing: " + apiKey, throttleTimeField);
+        }
     }
